# openapi.yaml
openapi: 3.0.3
info:
  title: Student API
  description: API สำหรับจัดการข้อมูลนิสิต พัฒนาในรายวิชา 88736065
  version: 1.0.0

servers:
  - url: http://localhost:3000/api/v1
    description: Development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string

    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "student@example.com"
        password:
          type: string
          format: password
          example: "Passw0rd!"

    RegisterResponse:
      type: object
      properties:
        message:
          type: string
          example: "สมัครสมาชิกสำเร็จ"
        data:
          type: object
          properties:
            id:
              type: integer
              example: 1
            email:
              type: string
              example: "student@example.com"
            role:
              type: string
              example: "student"

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "student@example.com"
        password:
          type: string
          format: password
          example: "Passw0rd!"

    LoginResponse:
      type: object
      properties:
        message:
          type: string
          example: "เข้าสู่ระบบสำเร็จ"
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

paths:
  /auth/register:
    post:
      summary: สมัครสมาชิก
      description: ลงทะเบียนผู้ใช้ใหม่
      tags:
        - Authentication

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"

      responses:
        "201":
          description: สมัครสมาชิกสำเร็จ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"

        "400":
          description: ข้อมูลไม่ครบถ้วน
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: "VALIDATION_ERROR"
                  message: "กรุณาระบุ email และ password"

        "409":
          description: Email มีอยู่ในระบบแล้ว
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: "DUPLICATE_EMAIL"
                  message: "อีเมลนี้มีอยู่ในระบบแล้ว"

  /auth/login:
    post:
      summary: เข้าสู่ระบบ
      description: ตรวจสอบ email และ password และคืน JWT
      tags:
        - Authentication

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"

      responses:
        "200":
          description: เข้าสู่ระบบสำเร็จ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"

        "400":
          description: ข้อมูลไม่ครบถ้วน
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: "VALIDATION_ERROR"
                  message: "กรุณาระบุ email และ password"

        "401":
          description: Email หรือ password ไม่ถูกต้อง
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: "INVALID_CREDENTIALS"
                  message: "อีเมลหรือรหัสผ่านไม่ถูกต้อง"

  /auth/me:
    get:
      summary: ดูข้อมูลผู้ใช้ปัจจุบัน
      description: ดึงข้อมูลผู้ใช้จาก JWT ที่ส่งมาใน Authorization header
      tags:
        - Authentication

      security:
        - bearerAuth: []

      responses:
        "200":
          description: สำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "สำเร็จ"
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      email:
                        type: string
                        example: "student@example.com"
                      role:
                        type: string
                        example: "student"

        "401":
          description: ไม่มีหรือไม่สามารถยืนยันตัวตนด้วย JWT ได้
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /students:
    get:
      summary: ดึงรายการนักศึกษาทั้งหมด
      description: ดึงข้อมูลนักศึกษาพร้อม Pagination และ Filter ตามสาขา
      tags:
        - Students

      parameters:
        - name: page
          in: query
          description: หมายเลขหน้าที่ต้องการ
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1

        - name: limit
          in: query
          description: จำนวนข้อมูลต่อหน้า
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
          example: 10

        - name: major
          in: query
          description: กรองข้อมูลตามสาขา
          required: false
          schema:
            type: string
          example: "Information Technology"

      responses:
        "200":
          description: ดึงข้อมูลนักศึกษาสำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "สำเร็จ"
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        name:
                          type: string
                          example: "สมชาย ใจดี"
                        major:
                          type: string
                          example: "Information Technology"
                        email:
                          type: string
                          example: "somchai@example.com"

                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 10
                      total:
                        type: integer
                        example: 25
                      totalPages:
                        type: integer
                        example: 3

    post:
      summary: เพิ่มข้อมูลนักศึกษา
      description: เพิ่มข้อมูลนักศึกษาใหม่
      tags:
        - Students

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - major
                - email
              properties:
                name:
                  type: string
                  example: "สมชาย ใจดี"
                major:
                  type: string
                  example: "Information Technology"
                email:
                  type: string
                  format: email
                  example: "somchai@example.com"

      responses:
        "201":
          description: เพิ่มข้อมูลสำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "เพิ่มข้อมูลสำเร็จ"
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: "สมชาย ใจดี"
                      major:
                        type: string
                        example: "Information Technology"
                      email:
                        type: string
                        example: "somchai@example.com"

        "400":
          description: ข้อมูลไม่ครบถ้วน
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: "VALIDATION_ERROR"
                  message: "กรุณาระบุข้อมูลให้ครบถ้วน"

        "409":
          description: Email มีอยู่ในระบบแล้ว
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: "DUPLICATE_EMAIL"
                  message: "อีเมลนี้มีอยู่ในระบบแล้ว"

  /students/{id}:
    put:
      summary: แก้ไขข้อมูลนักศึกษา
      description: แก้ไขชื่อและสาขาของนักศึกษา
      tags:
        - Students

      parameters:
        - name: id
          in: path
          description: รหัสนักศึกษา
          required: true
          schema:
            type: integer
          example: 1

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - major
              properties:
                name:
                  type: string
                  example: "สมชาย ใจดี"
                major:
                  type: string
                  example: "Information Technology"

      responses:
        "200":
          description: แก้ไขข้อมูลสำเร็จ

        "400":
          description: ข้อมูลไม่ครบถ้วน
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "404":
          description: ไม่พบข้อมูลนักศึกษา
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: "NOT_FOUND"
                  message: "ไม่พบข้อมูลนิสิต"

    delete:
      summary: ลบข้อมูลนักศึกษา
      description: ลบข้อมูลนักศึกษา โดยผู้ใช้ต้องมี role เป็น admin
      tags:
        - Students

      security:
        - bearerAuth: []

      parameters:
        - name: id
          in: path
          description: รหัสนักศึกษา
          required: true
          schema:
            type: integer
          example: 1

      responses:
        "200":
          description: ลบข้อมูลสำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "ลบข้อมูลสำเร็จ"

        "401":
          description: ไม่มีหรือไม่สามารถยืนยันตัวตนด้วย JWT ได้
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "403":
          description: ไม่มีสิทธิ์เข้าถึง เนื่องจาก role ไม่ใช่ admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: "FORBIDDEN"
                  message: "ไม่มีสิทธิ์เข้าถึง"

        "404":
          description: ไม่พบข้อมูลนักศึกษา
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: "NOT_FOUND"
                  message: "ไม่พบข้อมูลนิสิต"